refactor(codegen): start splitting codegen.rs into per-construct files
Item #2 from doc/plan_compiler_refactor.md, first installment. The
9.3k-line monolith is becoming a `codegen/` module directory mirroring
the layout `sim_codegen/` already uses (`mod.rs` + per-construct files
contributing `pub(super) fn emit_<name>` to `impl Codegen`).

This PR establishes the pattern with three constructs:

- `src/codegen/clkgate.rs` — 59 lines (smallest; quickest validation
  of the submodule wiring + visibility rules).
- `src/codegen/regfile.rs` — 334 lines including the
  `emit_regfile_port_scalar` helper (kept private to this submodule).
- `src/codegen/synchronizer.rs` — 326 lines.

Visibility note: each submodule does `use super::*;` and adds
`impl<'a> Codegen<'a> { pub(super) fn emit_<name> ... }`. Child modules
can access the parent's private struct fields (`indent`, `out`, etc.)
because Rust's privacy boundary lets descendant modules see private
items of the defining module — the same approach `sim_codegen/` uses.

Net result:
- `mod.rs`: 9183 → 8477 lines (−706, three constructs moved out).
- 3 new submodule files totaling 719 lines (effectively the moved
  bodies plus the per-file boilerplate header / `impl` wrapper).

Verified end-to-end:
- 221/221 integration tests pass.
- 24 lib unit tests pass.
- Zero snapshot drift across 34 SV-emission snapshots — the move is
  pure code-relocation, no behavior change.
- `cargo build` clean.

Follow-ups (queued, same pattern):
- counter, arbiter, ram, cam, fifo, fsm, pipeline, linklist, module.
- Each is independent; can land in any order. The biggest payoffs
  remaining are `module` (~3k lines), `fifo` (~1.4k), `pipeline`
  (~1.3k); the rest are 200–700 lines each.
