+++ src/codegen/mod.rs
@@ -4,6 +4,14 @@ use crate::lexer::Span;
 use crate::resolve::{Symbol, SymbolTable};
 use crate::typecheck::enum_width;
 
+// Per-construct submodules. Each contributes `pub(super) fn emit_<name>`
+// to `impl Codegen` and lives in its own file mirroring the layout of
+// `sim_codegen/`. New constructs land in their own file rather than
+// growing this `mod.rs`.
+mod clkgate;
+mod regfile;
+mod synchronizer;
+
 /// SV assignment-operator context for the unified `emit_stmt` walker.
 /// `Blocking` = `=` (comb / latch / reg-as-comb), `NonBlocking` = `<=` (seq).
 #[derive(Copy, Clone, PartialEq, Eq, Debug)]
@@ -6633,380 +6641,6 @@ impl<'a> Codegen<'a> {
     }
 
     // ── Synchronizer ─────────────────────────────────────────────────────────
-
-    fn emit_clkgate(&mut self, c: &crate::ast::ClkGateDecl) {
-        let n = &c.name.name;
-
-        // Find port names
-        let clk_in = c.ports.iter().find(|p| matches!(&p.ty, TypeExpr::Clock(_)) && p.direction == Direction::In)
-            .map(|p| p.name.name.as_str()).unwrap_or("clk_in");
-        let clk_out = c.ports.iter().find(|p| matches!(&p.ty, TypeExpr::Clock(_)) && p.direction == Direction::Out)
-            .map(|p| p.name.name.as_str()).unwrap_or("clk_out");
-        let enable = c.ports.iter().find(|p| p.name.name == "enable")
-            .map(|p| p.name.name.as_str()).unwrap_or("enable");
-        let test_en = c.ports.iter().find(|p| p.name.name == "test_en")
-            .map(|p| p.name.name.as_str());
-
-        // Module header
-        self.line(&format!("module {n} ("));
-        self.indent += 1;
-        let port_strs: Vec<String> = c.ports.iter().map(|p| {
-            let dir = match p.direction { Direction::In => "input", Direction::Out => "output" };
-            let ty = self.emit_port_type_str(&p.ty);
-            format!("{dir} {ty} {}", p.name.name)
-        }).collect();
-        for (i, ps) in port_strs.iter().enumerate() {
-            let comma = if i < port_strs.len() - 1 { "," } else { "" };
-            self.line(&format!("{ps}{comma}"));
-        }
-        self.indent -= 1;
-        self.line(");");
-        self.line("");
-        self.indent += 1;
-
-        let en_expr = if let Some(te) = test_en {
-            format!("{enable} | {te}")
-        } else {
-            enable.to_string()
-        };
-
-        match c.kind {
-            crate::ast::ClkGateKind::Latch => {
-                self.line("logic en_latched;");
-                self.line(&format!("always_latch if (!{clk_in}) en_latched = {en_expr};"));
-                self.line(&format!("assign {clk_out} = {clk_in} & en_latched;"));
-            }
-            crate::ast::ClkGateKind::And => {
-                self.line(&format!("assign {clk_out} = {clk_in} & ({en_expr});"));
-            }
-        }
-
-        self.indent -= 1;
-        self.line("");
-        self.line("endmodule");
-        self.line("");
-    }
-
-    fn emit_synchronizer(&mut self, s: &SynchronizerDecl) {
-        let n = &s.name.name;
-
-        // Resolve STAGES (default 2)
-        let stages = s.params.iter()
-            .find(|p| p.name.name == "STAGES")
-            .and_then(|p| p.default.as_ref())
-            .and_then(|e| if let ExprKind::Literal(LitKind::Dec(v)) = &e.kind { Some(*v as usize) } else { None })
-            .unwrap_or(2);
-
-        // Find clock ports (first = source clock, second = destination clock)
-        let clk_ports: Vec<&PortDecl> = s.ports.iter()
-            .filter(|p| matches!(&p.ty, TypeExpr::Clock(_)))
-            .collect();
-        let src_clk = &clk_ports[0].name.name;
-        let dst_clk = &clk_ports[1].name.name;
-
-        // Find data ports
-        let data_in_port = s.ports.iter().find(|p| p.name.name == "data_in").unwrap();
-        let data_ty = self.emit_port_type_str(&data_in_port.ty);
-
-        // Check for reset port
-        let rst_port = s.ports.iter().find(|p| matches!(&p.ty, TypeExpr::Reset(..)));
-
-        // Module header — emit all declared params as SV parameters
-        self.line(&format!("module {n} #("));
-        self.indent += 1;
-        let param_strs: Vec<String> = s.params.iter().map(|p| {
-            let val = p.default.as_ref()
-                .and_then(|e| if let ExprKind::Literal(LitKind::Dec(v)) = &e.kind { Some(v.to_string()) } else { None })
-                .unwrap_or_else(|| "0".to_string());
-            format!("parameter int {} = {}", p.name.name, val)
-        }).collect();
-        // Always include STAGES if not already declared
-        let has_stages = s.params.iter().any(|p| p.name.name == "STAGES");
-        let mut all_param_strs = param_strs;
-        if !has_stages {
-            all_param_strs.push(format!("parameter int STAGES = {stages}"));
-        }
-        for (i, ps) in all_param_strs.iter().enumerate() {
-            let comma = if i < all_param_strs.len() - 1 { "," } else { "" };
-            self.line(&format!("{ps}{comma}"));
-        }
-        self.indent -= 1;
-        self.line(") (");
-        self.indent += 1;
-        // Emit ports
-        let port_strs: Vec<String> = s.ports.iter().map(|p| {
-            let dir = match p.direction { Direction::In => "input", Direction::Out => "output" };
-            let ty = self.emit_port_type_str(&p.ty);
-            format!("{dir} {ty} {}", p.name.name)
-        }).collect();
-        for (i, ps) in port_strs.iter().enumerate() {
-            let comma = if i < port_strs.len() - 1 { "," } else { "" };
-            self.line(&format!("{ps}{comma}"));
-        }
-        self.indent -= 1;
-        self.line(");");
-        self.line("");
-        self.indent += 1;
-
-        match s.kind {
-            SyncKind::Ff => self.emit_sync_ff(dst_clk, &data_ty, rst_port, stages),
-            SyncKind::Gray => self.emit_sync_gray(src_clk, dst_clk, &data_ty, rst_port, stages),
-            SyncKind::Handshake => self.emit_sync_handshake(src_clk, dst_clk, &data_ty, rst_port, stages),
-            SyncKind::Reset => self.emit_sync_reset(dst_clk, rst_port, stages),
-            SyncKind::Pulse => self.emit_sync_pulse(src_clk, dst_clk, rst_port, stages),
-        }
-
-        self.indent -= 1;
-        self.line("");
-        self.line("endmodule");
-        self.line("");
-    }
-
-    // ── Synchronizer kind helpers ────────────────────────────────────────────
-
-    fn emit_sync_reset_begin(&mut self, dst_clk: &str, rst_port: Option<&PortDecl>) -> Option<String> {
-        if let Some(rp) = rst_port {
-            let is_low = matches!(&rp.ty, TypeExpr::Reset(_, level) if *level == ResetLevel::Low);
-            let is_async = matches!(&rp.ty, TypeExpr::Reset(sync_type, _) if *sync_type == ResetKind::Async);
-            let sensitivity = if is_async {
-                let edge = if is_low { "negedge" } else { "posedge" };
-                format!(" or {edge} {}", rp.name.name)
-            } else {
-                String::new()
-            };
-            self.line(&format!("always_ff @(posedge {dst_clk}{sensitivity}) begin"));
-            let cond = if is_low { format!("!{}", rp.name.name) } else { rp.name.name.clone() };
-            Some(cond)
-        } else {
-            self.line(&format!("always_ff @(posedge {dst_clk}) begin"));
-            None
-        }
-    }
-
-    fn emit_sync_ff(&mut self, dst_clk: &str, data_ty: &str, rst_port: Option<&PortDecl>, stages: usize) {
-        self.line(&format!("// {stages}-stage FF synchronizer chain (destination clock: {dst_clk})"));
-        self.line(&format!("{data_ty} sync_chain [0:STAGES-1];"));
-        self.line("");
-
-        let rst_cond = self.emit_sync_reset_begin(dst_clk, rst_port);
-        self.indent += 1;
-        if let Some(ref cond) = rst_cond {
-            self.line(&format!("if ({cond}) begin"));
-            self.indent += 1;
-            self.line("for (int i = 0; i < STAGES; i++) sync_chain[i] <= '0;");
-            self.indent -= 1;
-            self.line("end else begin");
-            self.indent += 1;
-        }
-        self.line("sync_chain[0] <= data_in;");
-        self.line("for (int i = 1; i < STAGES; i++) sync_chain[i] <= sync_chain[i-1];");
-        if rst_cond.is_some() {
-            self.indent -= 1;
-            self.line("end");
-        }
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-        self.line("assign data_out = sync_chain[STAGES-1];");
-    }
-
-    fn emit_sync_gray(&mut self, src_clk: &str, dst_clk: &str, data_ty: &str, rst_port: Option<&PortDecl>, stages: usize) {
-        self.line(&format!("// Gray-code synchronizer ({stages} stages, {src_clk} → {dst_clk})"));
-        self.line(&format!("{data_ty} bin_to_gray;"));
-        self.line(&format!("{data_ty} gray_chain [0:STAGES-1];"));
-        self.line(&format!("{data_ty} gray_to_bin;"));
-        self.line("");
-
-        // Binary-to-gray encode (combinational, source domain)
-        self.line("assign bin_to_gray = data_in ^ (data_in >> 1);");
-        self.line("");
-
-        // FF chain on destination clock
-        let rst_cond = self.emit_sync_reset_begin(dst_clk, rst_port);
-        self.indent += 1;
-        if let Some(ref cond) = rst_cond {
-            self.line(&format!("if ({cond}) begin"));
-            self.indent += 1;
-            self.line("for (int i = 0; i < STAGES; i++) gray_chain[i] <= '0;");
-            self.indent -= 1;
-            self.line("end else begin");
-            self.indent += 1;
-        }
-        self.line("gray_chain[0] <= bin_to_gray;");
-        self.line("for (int i = 1; i < STAGES; i++) gray_chain[i] <= gray_chain[i-1];");
-        if rst_cond.is_some() {
-            self.indent -= 1;
-            self.line("end");
-        }
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-
-        // Gray-to-binary decode: binary[i] = XOR of all gray bits from MSB down to i
-        // Computed as: b = g ^ (g>>1) ^ (g>>2) ^ ... — no self-reference, no ordering issue.
-        self.line("// Gray-to-binary decode (prefix XOR — no self-reference)");
-        self.line(&format!("always_comb begin"));
-        self.indent += 1;
-        self.line("gray_to_bin = gray_chain[STAGES-1];");
-        self.line(&format!("for (int i = 1; i < $bits({data_ty}); i++)"));
-        self.indent += 1;
-        self.line("gray_to_bin ^= gray_chain[STAGES-1] >> i;");
-        self.indent -= 1;
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-        self.line("assign data_out = gray_to_bin;");
-    }
-
-    fn emit_sync_handshake(&mut self, src_clk: &str, dst_clk: &str, data_ty: &str, rst_port: Option<&PortDecl>, stages: usize) {
-        self.line(&format!("// Handshake synchronizer ({stages} stages, {src_clk} → {dst_clk})"));
-        self.line(&format!("{data_ty} data_reg;"));
-        self.line("logic req_src, ack_src;");
-        self.line(&format!("logic req_sync [0:STAGES-1];  // req synchronized to {dst_clk}"));
-        self.line(&format!("logic ack_sync [0:STAGES-1];  // ack synchronized to {src_clk}"));
-        self.line("logic ack_dst;");
-        self.line("");
-
-        let rst_name = rst_port.map(|rp| rp.name.name.as_str()).unwrap_or("1'b0");
-        let is_low = rst_port.map_or(false, |rp| matches!(&rp.ty, TypeExpr::Reset(_, level) if *level == ResetLevel::Low));
-        let rst_active = if is_low { format!("!{rst_name}") } else { rst_name.to_string() };
-
-        // Source domain: latch data and toggle req
-        self.line(&format!("// Source domain ({src_clk}): latch data, manage req/ack"));
-        self.line(&format!("always_ff @(posedge {src_clk}) begin"));
-        self.indent += 1;
-        self.line(&format!("if ({rst_active}) begin"));
-        self.indent += 1;
-        self.line("req_src <= 1'b0;");
-        self.line("data_reg <= '0;");
-        self.indent -= 1;
-        self.line("end else if (data_in !== data_reg && req_src == ack_src) begin");
-        self.indent += 1;
-        self.line("data_reg <= data_in;");
-        self.line("req_src <= ~req_src;");
-        self.indent -= 1;
-        self.line("end");
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-
-        // Synchronize req into destination domain
-        self.line(&format!("// Synchronize req into {dst_clk}"));
-        self.line(&format!("always_ff @(posedge {dst_clk}) begin"));
-        self.indent += 1;
-        self.line(&format!("if ({rst_active}) begin"));
-        self.indent += 1;
-        self.line("for (int i = 0; i < STAGES; i++) req_sync[i] <= 1'b0;");
-        self.line("ack_dst <= 1'b0;");
-        self.indent -= 1;
-        self.line("end else begin");
-        self.indent += 1;
-        self.line("req_sync[0] <= req_src;");
-        self.line("for (int i = 1; i < STAGES; i++) req_sync[i] <= req_sync[i-1];");
-        self.line("ack_dst <= req_sync[STAGES-1];");
-        self.indent -= 1;
-        self.line("end");
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-
-        // Synchronize ack back into source domain
-        self.line(&format!("// Synchronize ack back into {src_clk}"));
-        self.line(&format!("always_ff @(posedge {src_clk}) begin"));
-        self.indent += 1;
-        self.line(&format!("if ({rst_active}) begin"));
-        self.indent += 1;
-        self.line("for (int i = 0; i < STAGES; i++) ack_sync[i] <= 1'b0;");
-        self.indent -= 1;
-        self.line("end else begin");
-        self.indent += 1;
-        self.line("ack_sync[0] <= ack_dst;");
-        self.line("for (int i = 1; i < STAGES; i++) ack_sync[i] <= ack_sync[i-1];");
-        self.indent -= 1;
-        self.line("end");
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-
-        self.line("assign ack_src = ack_sync[STAGES-1];");
-        self.line("assign data_out = data_reg;");
-    }
-
-    fn emit_sync_reset(&mut self, dst_clk: &str, _rst_port: Option<&PortDecl>, _stages: usize) {
-        // Reset synchronizer: data_in is the async reset input (active high).
-        // Assert immediately (async), deassert through N-stage FF chain (sync to dst_clk).
-        self.line(&format!("// Reset synchronizer: async assert, sync deassert on {dst_clk}"));
-        self.line("logic sync_chain [0:STAGES-1];");
-        self.line("");
-
-        self.line(&format!("always_ff @(posedge {dst_clk} or posedge data_in) begin"));
-        self.indent += 1;
-        self.line("if (data_in) begin");
-        self.indent += 1;
-        self.line("for (int i = 0; i < STAGES; i++) sync_chain[i] <= 1'b1;");
-        self.indent -= 1;
-        self.line("end else begin");
-        self.indent += 1;
-        self.line("sync_chain[0] <= 1'b0;");
-        self.line("for (int i = 1; i < STAGES; i++) sync_chain[i] <= sync_chain[i-1];");
-        self.indent -= 1;
-        self.line("end");
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-        self.line("assign data_out = sync_chain[STAGES-1];");
-    }
-
-    fn emit_sync_pulse(&mut self, src_clk: &str, dst_clk: &str, rst_port: Option<&PortDecl>, _stages: usize) {
-        let rst_name = rst_port.map(|rp| rp.name.name.as_str());
-        let is_low = rst_port.map_or(false, |rp| matches!(&rp.ty, TypeExpr::Reset(_, level) if *level == ResetLevel::Low));
-        let rst_cond = rst_name.map(|n| if is_low { format!("!{n}") } else { n.to_string() });
-
-        self.line(&format!("// Pulse synchronizer: {src_clk} → {dst_clk}"));
-        self.line("// Source: pulse → toggle; Destination: sync toggle → edge detect → pulse");
-        self.line("logic toggle_src;");
-        self.line("logic sync_chain [0:STAGES-1];");
-        self.line("logic pulse_dst;");
-        self.line("");
-
-        // Source domain: toggle on input pulse
-        self.line(&format!("always_ff @(posedge {src_clk}) begin"));
-        self.indent += 1;
-        if let Some(ref cond) = rst_cond {
-            self.line(&format!("if ({cond}) toggle_src <= 1'b0;"));
-            self.line("else if (data_in) toggle_src <= ~toggle_src;");
-        } else {
-            self.line("if (data_in) toggle_src <= ~toggle_src;");
-        }
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-
-        // Destination domain: sync the toggle through FF chain
-        self.line(&format!("always_ff @(posedge {dst_clk}) begin"));
-        self.indent += 1;
-        if let Some(ref cond) = rst_cond {
-            self.line(&format!("if ({cond}) begin"));
-            self.indent += 1;
-            self.line("for (int i = 0; i < STAGES; i++) sync_chain[i] <= 1'b0;");
-            self.indent -= 1;
-            self.line("end else begin");
-            self.indent += 1;
-        }
-        self.line("sync_chain[0] <= toggle_src;");
-        self.line("for (int i = 1; i < STAGES; i++) sync_chain[i] <= sync_chain[i-1];");
-        if rst_cond.is_some() {
-            self.indent -= 1;
-            self.line("end");
-        }
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-
-        // Edge detect: XOR of last two stages → single-cycle pulse
-        self.line("assign data_out = sync_chain[STAGES-1] ^ sync_chain[STAGES-2];");
-    }
-
     // ── RAM ───────────────────────────────────────────────────────────────────
 
     fn emit_ram(&mut self, r: &RamDecl) {
@@ -7944,329 +7578,6 @@ impl<'a> Codegen<'a> {
         self.line("end");
     }
 
-    // ── Regfile ───────────────────────────────────────────────────────────────
-
-    fn emit_regfile(&mut self, r: &crate::ast::RegfileDecl) {
-        use crate::ast::ParamKind;
-        let n = &r.name.name.clone();
-
-        let nregs = r.param_int("NREGS", 32);
-
-        // Data width: prefer the UInt<N> type of the data signal in the write port,
-        // then fall back to XLEN/WIDTH/DATA_WIDTH params.
-        let data_width_num: String = r.write_ports.as_ref()
-            .and_then(|wp| wp.signals.iter().find(|s| s.name.name == "data"))
-            .and_then(|s| if let TypeExpr::UInt(w) = &s.ty { Some(self.emit_expr_str(w)) } else { None })
-            .or_else(|| {
-                r.params.iter()
-                    .find(|p| matches!(p.name.name.as_str(), "XLEN" | "WIDTH" | "DATA_WIDTH"))
-                    .and_then(|p| match &p.kind {
-                        ParamKind::Const | ParamKind::WidthConst(..) => p.default.as_ref().map(|e| self.emit_expr_str(e)),
-                        _ => None,
-                    })
-            })
-            .unwrap_or_else(|| "32".to_string());
-
-        // Determine addr width: ceil(log2(NREGS))
-        let addr_width = crate::width::index_width(nregs);
-
-        // Read/write port counts — resolve param references
-        let nread = r.read_ports.as_ref()
-            .map(|rp| r.resolve_count_expr(&rp.count_expr))
-            .unwrap_or(1);
-        let nwrite = r.write_ports.as_ref()
-            .map(|wp| r.resolve_count_expr(&wp.count_expr))
-            .unwrap_or(1);
-
-        let clk = r.ports.iter()
-            .find(|p| matches!(&p.ty, TypeExpr::Clock(_)))
-            .map(|p| p.name.name.clone())
-            .unwrap_or_else(|| "clk".to_string());
-        let (rst, is_async, is_low) = Self::extract_reset_info(&r.ports);
-
-        // ── Module header ─────────────────────────────────────────────────────
-        // Emit one SV parameter per ARCH param declaration
-        self.line(&format!("module {n} #("));
-        self.indent += 1;
-        let param_count = r.params.len();
-        for (i, p) in r.params.iter().enumerate() {
-            let comma = if i < param_count - 1 { "," } else { "" };
-            let val = p.default.as_ref().map(|e| self.emit_expr_str(e)).unwrap_or_else(|| "0".to_string());
-            self.line(&format!("parameter int {} = {}{}", p.name.name, val, comma));
-        }
-        self.indent -= 1;
-        self.line(") (");
-        self.indent += 1;
-
-        // Build port list — multi-port groups are flattened to scalar ports
-        // named {pfx}{i}_{signal} (e.g. read0_addr, read1_addr) so that each
-        // can be connected individually via a `connect` statement.
-        let mut all_ports: Vec<String> = Vec::new();
-        for p in &r.ports {
-            let dir = match p.direction { Direction::In => "input", Direction::Out => "output" };
-            let ty_str = self.emit_port_type_str(&p.ty);
-            all_ports.push(format!("{dir} {ty_str} {}", p.name.name));
-        }
-        // Read ports — one set of scalars per port index
-        if let Some(rp) = &r.read_ports {
-            let pfx = &rp.name.name;
-            for i in 0..nread {
-                for s in &rp.signals {
-                    let dir = match s.direction { Direction::In => "input", Direction::Out => "output" };
-                    let flat_name = if nread == 1 {
-                        format!("{pfx}_{}", s.name.name)
-                    } else {
-                        format!("{pfx}{i}_{}", s.name.name)
-                    };
-                    let port_str = self.emit_regfile_port_scalar(dir, &s.ty, &flat_name, addr_width, &data_width_num);
-                    all_ports.push(port_str);
-                }
-            }
-        }
-        // Write ports — one set of scalars per port index
-        if let Some(wp) = &r.write_ports {
-            let pfx = &wp.name.name;
-            for i in 0..nwrite {
-                for s in &wp.signals {
-                    let dir = match s.direction { Direction::In => "input", Direction::Out => "output" };
-                    let flat_name = if nwrite == 1 {
-                        format!("{pfx}_{}", s.name.name)
-                    } else {
-                        format!("{pfx}{i}_{}", s.name.name)
-                    };
-                    let port_str = self.emit_regfile_port_scalar(dir, &s.ty, &flat_name, addr_width, &data_width_num);
-                    all_ports.push(port_str);
-                }
-            }
-        }
-        let port_count = all_ports.len();
-        for (i, p) in all_ports.iter().enumerate() {
-            let comma = if i < port_count - 1 { "," } else { "" };
-            self.line(&format!("{p}{comma}"));
-        }
-        self.indent -= 1;
-        self.line(");");
-        self.line("");
-        self.indent += 1;
-
-        // ── Register array ────────────────────────────────────────────────────
-        self.line(&format!("logic [{}] rf_data [0:NREGS-1];", Self::fold_width_str(&data_width_num)));
-        self.line("");
-
-        // ── Determine read/write port signal names (flat) ─────────────────────
-        let write_pfx = r.write_ports.as_ref().map(|wp| wp.name.name.clone()).unwrap_or_else(|| "write".to_string());
-        let read_pfx  = r.read_ports.as_ref().map(|rp| rp.name.name.clone()).unwrap_or_else(|| "read".to_string());
-
-        // Flat name helper: "{pfx}{i}_{sig}" when count>1, else "{pfx}_{sig}"
-        let flat = |pfx: &str, i: u64, count: u64, sig: &str| -> String {
-            if count == 1 { format!("{pfx}_{sig}") } else { format!("{pfx}{i}_{sig}") }
-        };
-
-        // ── Write storage ─────────────────────────────────────────────────────
-        // Collect init-guarded addresses: init[k]=v means addr k is immutable
-        // (implemented as a write guard), not as a reset.
-        let guarded_addrs: Vec<String> = r.inits.iter()
-            .map(|init| self.emit_expr_str(&init.index))
-            .collect();
-
-        // Latch storage path — emit per-row `always_latch` with one-hot
-        // write decoding. Saves area / power on ASIC; FPGA tools will flag
-        // this at mapping (intentional — ARCH is target-agnostic).
-        if r.kind == crate::ast::RegfileKind::Latch {
-            let wen_in   = flat(&write_pfx, 0, nwrite, "en");
-            let waddr_in = flat(&write_pfx, 0, nwrite, "addr");
-            let wdata_in = flat(&write_pfx, 0, nwrite, "data");
-
-            // `flops: internal` (Ibex-style) — regfile auto-emits its own
-            // sample flops + per-row ICG-equivalent gating, so the caller can
-            // drive write pins combinationally. Adds 1-cycle write latency:
-            // a write asserted in cycle N is captured into rf_data[k] during
-            // cycle N+1's clk-low phase. The flop->latch pipeline keeps
-            // wdata stable across the latch transparency window without any
-            // contract on the caller.
-            let internal_flops = matches!(r.flops, crate::ast::RegfileFlops::Internal);
-            let (wen_eff, waddr_eff, wdata_eff) = if internal_flops {
-                let aw_msb = addr_width.saturating_sub(1);
-                let dw = Self::fold_width_str(&data_width_num);
-                self.line("// Internal sample flops (regfile flops: internal)");
-                self.line("logic                 we_q;");
-                self.line(&format!("logic [{aw_msb}:0]        waddr_q;"));
-                self.line(&format!("logic [{dw}]            wdata_q;"));
-                self.line(&format!("always_ff @(posedge {clk}) begin"));
-                self.indent += 1;
-                self.line(&format!("we_q <= {wen_in};"));
-                self.line(&format!("if ({wen_in}) begin"));
-                self.indent += 1;
-                self.line(&format!("waddr_q <= {waddr_in};"));
-                self.line(&format!("wdata_q <= {wdata_in};"));
-                self.indent -= 1;
-                self.line("end");
-                self.indent -= 1;
-                self.line("end");
-                self.line("");
-                ("we_q".to_string(), "waddr_q".to_string(), "wdata_q".to_string())
-            } else {
-                (wen_in, waddr_in, wdata_in)
-            };
-
-            for k in 0..nregs {
-                let k_lit = format!("{addr_width}'d{k}");
-                let init_for_k = r.inits.iter()
-                    .find_map(|init| match &init.index.kind {
-                        crate::ast::ExprKind::Literal(crate::ast::LitKind::Dec(v)) if *v == k => {
-                            Some(self.emit_expr_str(&init.value))
-                        }
-                        _ => None,
-                    });
-                if let Some(val) = init_for_k {
-                    // Address with `init[k]=v` is immutable — drive a constant.
-                    self.line(&format!("assign rf_data[{k}] = {val};"));
-                } else {
-                    self.line(&format!("always_latch begin"));
-                    self.indent += 1;
-                    if internal_flops {
-                        // ICG-equivalent gating: latch transparent only during
-                        // clk-low of the cycle after the sample, when we_q /
-                        // waddr_q / wdata_q are stable.
-                        self.line(&format!("if (!{clk} && {wen_eff} && {waddr_eff} == {k_lit})"));
-                    } else {
-                        self.line(&format!("if ({wen_eff} && {waddr_eff} == {k_lit})"));
-                    }
-                    self.indent += 1;
-                    self.line(&format!("rf_data[{k}] = {wdata_eff};"));
-                    self.indent -= 1;
-                    self.indent -= 1;
-                    self.line("end");
-                }
-            }
-            self.line("");
-        } else {
-
-        // Only include reset sensitivity when a reset port is actually present
-        // and there are reset-driven init entries. For register files that use
-        // write guards for x0 (not reset), emit plain posedge-only always_ff.
-        let has_reset_port = !rst.is_empty();
-        let use_reset = has_reset_port && r.inits.iter().any(|_| false); // reserved for future explicit reset-on-init
-
-        let ff_sens = if use_reset {
-            Self::ff_sensitivity(&clk, &rst, is_async, is_low)
-        } else {
-            format!("posedge {clk}")
-        };
-
-        self.line(&format!("always_ff @({ff_sens}) begin"));
-        self.indent += 1;
-
-        if use_reset {
-            let rst_cond = Self::rst_condition(&rst, is_low);
-            self.line(&format!("if ({rst_cond}) begin"));
-            self.indent += 1;
-            for init in &r.inits {
-                let idx = self.emit_expr_str(&init.index);
-                let val = self.emit_expr_str(&init.value);
-                self.line(&format!("rf_data[{idx}] <= {val};"));
-            }
-            self.indent -= 1;
-            self.line("end else begin");
-            self.indent += 1;
-        }
-
-        // Unroll write ports; add address guards for init[k] entries
-        for wi in 0..nwrite {
-            let wen   = flat(&write_pfx, wi, nwrite, "en");
-            let waddr = flat(&write_pfx, wi, nwrite, "addr");
-            let wdata = flat(&write_pfx, wi, nwrite, "data");
-            // Build guard: skip writes to any init-protected address
-            let addr_guards: Vec<String> = guarded_addrs.iter()
-                .map(|a| format!("{waddr} != {a}"))
-                .collect();
-            let guard = if addr_guards.is_empty() {
-                wen.clone()
-            } else {
-                format!("{wen} && {}", addr_guards.join(" && "))
-            };
-            self.line(&format!("if ({guard})"));
-            self.indent += 1;
-            self.line(&format!("rf_data[{waddr}] <= {wdata};"));
-            self.indent -= 1;
-        }
-
-        if use_reset {
-            self.indent -= 1;
-            self.line("end");
-        }
-
-        self.indent -= 1;
-        self.line("end");
-        self.line("");
-        }  // end flop branch (else of `kind == Latch`)
-
-        // ── Read always_comb — unrolled per read port ─────────────────────────
-        self.line("always_comb begin");
-        self.indent += 1;
-        for ri in 0..nread {
-            let raddr = flat(&read_pfx, ri, nread, "addr");
-            let rdata = flat(&read_pfx, ri, nread, "data");
-            if r.forward_write_before_read {
-                // Forward from the first write port (write port 0)
-                let wen   = flat(&write_pfx, 0, nwrite, "en");
-                let waddr = flat(&write_pfx, 0, nwrite, "addr");
-                let wdata = flat(&write_pfx, 0, nwrite, "data");
-                self.line(&format!("if ({wen} && {waddr} == {raddr})"));
-                self.indent += 1;
-                self.line(&format!("{rdata} = {wdata};"));
-                self.indent -= 1;
-                self.line("else");
-                self.indent += 1;
-                self.line(&format!("{rdata} = rf_data[{raddr}];"));
-                self.indent -= 1;
-            } else {
-                self.line(&format!("{rdata} = rf_data[{raddr}];"));
-            }
-        }
-        self.indent -= 1;
-        self.line("end");
-
-        if !r.asserts.is_empty() {
-            let clk = r.ports.iter().find(|p| matches!(&p.ty, TypeExpr::Clock(_)))
-                .map(|p| p.name.name.clone()).unwrap_or_else(|| "clk".to_string());
-            self.line("");
-            let asserts = r.asserts.clone();
-            let rname = r.name.name.clone();
-            self.emit_asserts_for_construct(&asserts, &rname, &clk);
-        }
-
-        self.indent -= 1;
-        self.line("");
-        self.line("endmodule");
-        self.line("");
-    }
-
-    /// Emit a single scalar regfile port signal declaration.
-    fn emit_regfile_port_scalar(
-        &self,
-        dir: &str,
-        ty: &TypeExpr,
-        name: &str,
-        addr_width: u32,
-        data_width: &str,
-    ) -> String {
-        let phy_ty = match ty {
-            TypeExpr::Bool => "logic".to_string(),
-            TypeExpr::Named(id) if id.name == "WIDTH" || id.name == "DATA_WIDTH" => {
-                format!("logic [{}]", Self::fold_width_str(data_width))
-            }
-            TypeExpr::Named(id) if id.name == "ADDR_WIDTH" || id.name.to_lowercase().contains("addr") => {
-                format!("logic [{}:0]", addr_width - 1)
-            }
-            TypeExpr::UInt(w) => {
-                let ws = self.emit_expr_str(w);
-                format!("logic [{}]", Self::fold_width_str(&ws))
-            }
-            _ => self.emit_port_type_str(ty),
-        };
-        format!("{dir} {phy_ty} {name}")
-    }
 
     /// Map signal types: Named("WIDTH") → `logic [DATA_WIDTH-1:0]`; others pass through
     fn emit_ram_signal_type(&self, ty: &TypeExpr) -> String {
